refactor: split resolveTacticDecisionCommand into modular handlers
Split the 533-line monolithic file into a modular structure with one
handler per tactic type:
- rethink.ts (Day 2)
- sparingPower.ts (Night 6)
- manaSteal.ts (Day 3)
- preparation.ts (Night 5)
- midnightMeditation.ts (Night 4)

Also:
- Extract shared calculateTurnOrder to helpers.ts (was duplicated)
- Extract phase transition logic to helpers.ts
- Main command is now a thin dispatcher (~155 lines)
- Each handler is independently testable (~60-90 lines each)
