+++ packages/core/src/engine/__tests__/tactics.test.ts
@@ -24,7 +24,7 @@ import {
 } from "@mage-knight/shared";
 import { createTacticsSelectionState } from "./testHelpers.js";
 import { createSelectTacticCommand } from "../commands/selectTacticCommand.js";
-import { createResolveTacticDecisionCommand } from "../commands/resolveTacticDecisionCommand.js";
+import { createResolveTacticDecisionCommand } from "../commands/tactics/index.js";
 import { getTacticCard } from "../../data/tactics.js";
 import { getTurnOptions } from "../validActions/turn.js";
 import { getValidActions } from "../validActions/index.js";
